improvement(Payroll Modules) - Improve the way to send selected employee, send a array who contains the   reference of selected Employees - Deleting the request that runs in a loop, - Implementation of function commitment who manage Data that will be used   to execute the transaction to support the employees
