+++ server/controllers/payroll/multiplePayroll/find.js
@@ -14,20 +14,26 @@ function find(options) {
   // ensure epected options are parsed appropriately as binary
   const filters = new FilterParser(options, { tableAlias : 'payroll' });
   let statusIds = [];
+  let referenceEmployees = [];
 
   if (options.status_id) {
     statusIds = statusIds.concat(options.status_id);
   }
 
+  if (options.reference) {
+    referenceEmployees = options.reference;
+  }
+
+
   const sql = `
-    SELECT payroll.employee_uuid, payroll.code, payroll.date_embauche, payroll.nb_enfant, payroll.individual_salary,
+    SELECT payroll.employee_uuid, payroll.reference, payroll.code, payroll.date_embauche, payroll.nb_enfant, payroll.individual_salary,
      payroll.account_id, payroll.creditor_uuid, payroll.display_name, payroll.sex, payroll.uuid, 
      payroll.payroll_configuration_id, payroll.currency_id, payroll.paiement_date, payroll.base_taxable, 
      payroll.basic_salary, payroll.gross_salary, payroll.grade_salary, payroll.text, payroll.net_salary, 
      payroll.working_day, payroll.total_day, payroll.daily_salary, payroll.amount_paid,
       payroll.status_id, payroll.status, (payroll.net_salary - payroll.amount_paid) AS balance
     FROM(
-      SELECT BUID(employee.uuid) AS employee_uuid, employee.code, employee.date_embauche, employee.nb_enfant, 
+      SELECT BUID(employee.uuid) AS employee_uuid, employee.reference, employee.code, employee.date_embauche, employee.nb_enfant, 
       employee.individual_salary, creditor_group.account_id, BUID(employee.creditor_uuid) AS creditor_uuid,
         UPPER(patient.display_name) AS display_name, patient.sex, BUID(paiement.uuid) AS uuid, 
         paiement.payroll_configuration_id,  paiement.currency_id, paiement.paiement_date, paiement.base_taxable, 
@@ -46,7 +52,7 @@ function find(options) {
         JOIN paiement_status ON paiement_status.id = paiement.status_id
         WHERE paiement.payroll_configuration_id = '${options.payroll_configuration_id}'
       UNION 
-        SELECT BUID(employee.uuid) AS employee_uuid, employee.code, employee.date_embauche, employee.nb_enfant, 
+        SELECT BUID(employee.uuid) AS employee_uuid, employee.reference, employee.code, employee.date_embauche, employee.nb_enfant, 
         employee.individual_salary, creditor_group.account_id, BUID(employee.creditor_uuid) AS creditor_uuid,
         UPPER(patient.display_name) AS display_name, patient.sex, NULL AS 'paiement_uuid',
         '${options.payroll_configuration_id}' AS payroll_configuration_id, '${options.currency_id}' AS currency_id, 
@@ -78,6 +84,7 @@ function find(options) {
   }
 
   filters.custom('status_id', 'payroll.status_id IN (?)', [statusIds]);
+  filters.custom('reference', 'payroll.reference IN (?)', [referenceEmployees]);
   filters.setOrder('ORDER BY payroll.display_name');
 
   const query = filters.applyQuery(sql);
